feat: minor refactor (#2258)
