+++ next/content/lung/atlases/lung/v1.0/inclusion.mdx
