+++ next/content/nervous-system/atlases/brain/index.ts
