+++ next/content/nervous-system/atlases/brain/v1.0/description.mdx
