+++ next/content/nervous-system/atlases/brain/v1.0/inclusion.mdx
