+++ next/content/nervous-system/atlases/cortex/index.ts
