+++ next/content/nervous-system/atlases/cortex/v1.0/description.mdx
